Make minimal changes to eliminate errors
In particular, it is a runtime error to raise StopIteration directly inside a generator function, see the following PEP:

https://peps.python.org/pep-0479

Also, assigning lambdas to variables is so frowned upon that VSCode keeps fixing it every time I touch that cell, even with format-on-save turned off.  I have reworked the example to prevent the possibility that the fix will be committed accidentally and make the example nonsensical.